media/sample: UampSettingsScreen M3 cleanup (TimeText, minimumVerticalContentPadding, pure M3 imports)
Review feedback:
- Previews now wrap the screen in M3 AppScaffold + TimeText with a
  fixed 10:10 TimeSource, so system UI is visible in rendered PNGs.
- Each TLC item applies Modifier.minimumVerticalContentPadding with
  the M3-recommended constants (ListHeaderDefaults.minimumTopListContentPadding
  for the header, ButtonDefaults.minimumVerticalListContentPadding for
  buttons/checkbox).
- Move the M2 ActionSetting / CheckedSetting helpers out to
  LegacySettingHelpers.kt so UampSettingsScreen.kt no longer imports
  androidx.wear.compose.material.*. The helpers stay M2 because
  DeveloperOptionsScreen / SamplesScreen still call them; both migrate
  (and the helpers delete) in follow-up commits on this branch.
